+++ P4/ParticleGravity.cpp
@@ -2,7 +2,7 @@
 
 ParticleGravity::~ParticleGravity() {}
 
-void ParticleGravity::updateForce(Particle * p, float t) {
+void ParticleGravity::updateForce(Particle * p, double t) {
 	if (!p->hasInfiniteMass()) return;
 
 	p->addForce(g * p->getMass());
